% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{bayes.ridge.mcmc}
\alias{bayes.ridge.mcmc}
\title{Metropolis-Hastings Sampler}
\usage{
bayes.ridge.mcmc(
  formula,
  data,
  chain_length = 60000,
  lambda_adapt = TRUE,
  seed = NULL,
  beta_init = NULL,
  lambda_init = NULL,
  trace = FALSE
)
}
\arguments{
\item{formula}{the formula to be used in the logistic regression model}

\item{data}{the complete data including response and covariates}

\item{chain_length}{the number of posterior draws to sample}

\item{lambda_adapt}{logical, include lambda in the adaptive proposal (TRUE) or generate independently (FALSE)}

\item{seed}{the seed to begin the sampler, for reproducability}

\item{beta_init}{the starting values for the regression coefficients}

\item{lambda_init}{the starting value for the shrinkage parameter lambda}

\item{trace}{logical, if TRUE show details while running sampler}
}
\value{
A list containing the following:
\item{chain}{The MCMC chain with the unstandardized coefficients}
\item{chain_std}{The MCMC with the standardized coefficients}
\item{acceptance_rate}{The acceptance rate for the sampler}
\item{design}{The design matrix}
\item{design_std}{The standardized design matrix used for the sampler}
}
\description{
This function generates an MCMC chain using adaptive Metropolis-Hastings
algorithm for a logistic regression problem, assuming i.i.d N(0, lambda^2) priors
on the regression coefficients conditional on lambda, and a standard half-Cauchy
prior on lambda.
}
