% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{cov_state}
\alias{cov_state}
\title{Covariance Estimate for Adaptive Metropolis-Hastings}
\usage{
cov_state(chain_mat, state, start)
}
\arguments{
\item{chain_mat}{the complete MCMC}

\item{state}{the index of the row of the chain corresponding to the current iteration}

\item{start}{the initial index of the chain matrix to compute the covariance}
}
\value{
the logarithm of a standard half-Cauchy density at lambda
}
\description{
This function computes the estimated covariance matrix at a given iteration of
the Metropolis-Hastings algorithm. A small diagonal matrix is added to ensure
positive-definiteness
}
