% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvPenalizedLogit.R
\name{cv.penalized.logit}
\alias{cv.penalized.logit}
\title{Cross-validated penalized logistic regression}
\usage{
cv.penalized.logit(X, y, n_folds = 5, lambda = NULL)
}
\arguments{
\item{X}{Matrix of predictors.}

\item{y}{Vector of response variables.}

\item{n_folds}{Number of folds for cross-validation.}

\item{lambda}{Vector of lambda values for penalization. If NULL, lambda values are explored automatically.}
}
\value{
A list containing:
\describe{
\item{\code{lambda.min}}{The lambda value corresponding to the minimum misclassification rate (mcr).}
\item{\code{se.mcr}}{Standard error of the misclassification rate (mcr).}
\item{\code{se.lambda.min}}{Standard error of the minimum mcr lambda value.}
\item{\code{se.1.lambda}}{Lambda value with mcr within 1 standard error of the minimum mcr.}
\item{\code{cv.out}}{Data frame containing lambda, log_lambda, and mcr values for each lambda.}
}
}
\description{
This function performs cross-validated penalized logistic regression using glmnet for.
minimizing the error rate
Taken and adapted from: www.r-bloggers.com/2021/10/lambda-min-lambda-1se-and-cross-validation-in-lasso-binomial-response/
}
\examples{
X <- matrix(rnorm(100), ncol = 5)
y <- sample(0:1, 20, replace = TRUE)
cv_result <- cv.penalized.logit(X, y)

}
