% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log-Likelihood for Logistic Regression}
\usage{
log_likelihood(y, x, beta)
}
\arguments{
\item{y}{the vector of responses}

\item{x}{the design matrix, including the intercept}

\item{beta}{specified regression coefficients}
}
\value{
the logarithm of a standard half-Cauchy density at lambda
}
\description{
This function computes the log-likelihood at a given beta vector
for a given response vector y and design matrix x.
}
