% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{log_pbeta}
\alias{log_pbeta}
\title{Logarithm of Normal Prior}
\usage{
log_pbeta(beta, lambda)
}
\arguments{
\item{beta}{a vector of numbers}

\item{lambda}{the prior sd}
}
\value{
the sum of the logarithm of the density at each beta value
}
\description{
This function computes the log of a normal density with mean 0 and
standard deviation lambda. It uses a noninformative prior on the first
parameter, assumed to be the intercept.
}
