% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{log_plambda}
\alias{log_plambda}
\title{Logarithm of Half-Cauchy Prior}
\usage{
log_plambda(lambda)
}
\arguments{
\item{lambda}{a number}
}
\value{
the logarithm of a standard half-Cauchy density at lambda
}
\description{
This function computes the logarithm of a standard half-Cauchy density.
}
