% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{log_posterior_unnormalized}
\alias{log_posterior_unnormalized}
\title{Unnormalized Log-Posterior}
\usage{
log_posterior_unnormalized(y, x, beta, lambda)
}
\arguments{
\item{y}{the vector of responses}

\item{x}{the design matrix, including the intercept}

\item{beta}{specified regression coefficients}

\item{lambda}{the standard deviation for the normal beta prior}
}
\value{
the logarithm of the log-posterior for logistic regression
}
\description{
This function computes the unnormalized log-posterior for logistic regression
assuming a normal prior on the betas and a standard half-Cauchy prior on lambda,
the standard deviation of the betas
}
