% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{Mode}
\alias{Mode}
\title{Mode of a distribution}
\usage{
Mode(x, adjust = 1)
}
\arguments{
\item{x}{the set of values with which to compute the mode}

\item{adjust}{bandwidth used for estimating the density}
}
\value{
the estimated mode
}
\description{
This function computes the mode of a set of numbers by approximating the density
}
