% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized.R
\name{penalized.logit}
\alias{penalized.logit}
\title{Penalized Logistic Regression}
\usage{
penalized.logit(X, y, beta, lambda, iter = 1000, tol = 1e-04)
}
\arguments{
\item{X}{Matrix of predictor variables.}

\item{y}{Vector of response variable.}

\item{beta}{Initial values of the coefficients.}

\item{lambda}{Vector of penalty parameters.}

\item{iter}{Maximum number of iterations (default: 1000).}

\item{tol}{Tolerance for convergence (default: 0.0001).}
}
\value{
A matrix containing the estimated coefficients for each lambda value.
}
\description{
This function performs penalized logistic regression using iterative updates
}
\details{
The function iteratively updates the coefficients using a penalized
logistic regression algorithm until convergence or maximum iterations are reached.
}
\examples{
data("diabetes_train")
predictors <- scale(model.matrix(diabetes ~ Age + Gender + BMI + SBP + DBP + height + 
                        Triglyceride + ALT + AST + BUN + CCR + Cholesterol +
                        factor(smoke) + factor(drink) + factor(history), data = diabetes_train)[,-1]
response <- diabetes_train$diabetes

# We get the best lambda
lambda_seq <- seq(0.013, 0.0008, by = -0.0001)
cv_lambda <- cv.penalized.logit(predictors, response, n_folds = 5, lambda = lambda_seq)
cv_lambda$se.1.lambda

# Penalized logistic regression solution
beta0 = glm(y~X, family = "binomial")$coefficients
penalized_logit_result <- penalized.logit(predictors, response, beta = beta0, lambda = cv_lambda$se.1.lambda) 
penalized_logit_result

}
