% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{proposal}
\alias{proposal}
\title{Proposal Function for Metropolis Hastings Random Walk}
\usage{
proposal(chain_mat, state, start, lambda_adapt = TRUE)
}
\arguments{
\item{chain_mat}{the complete MCMC}

\item{state}{the index of the row of the chain corresponding to the current iteration}

\item{start}{the initial index of the chain matrix to compute the covariance}

\item{lambda_adapt}{logical, include lambda in the adaptive proposal (TRUE) or generate independently (FALSE)}
}
\value{
a random draw from a multivariate normal density with covariance estimated from chain
}
\description{
This function generates a beta based on the estimated covariance matrix of the current chain.
It also generates a lambda proposal by eith including it in the covariance
matrix or generating it independently. Finally, the function returns the
concatenated c(beta, lambda) vector.
}
