% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_wrappers.R
\name{R}
\alias{R}
\title{Metropolis-Hastings Ratio}
\usage{
R(y, x, log_posterior_curr, log_posterior_prop)
}
\arguments{
\item{y}{the vector of responses}

\item{x}{the design matrix, including the intercept}

\item{log_posterior_curr}{the unnormalized current log posterior}

\item{log_posterior_prop}{the unnormalized proposal log posterior}
}
\value{
the minimum of 1 and the ratio of unnormalized posteriors
}
\description{
This function computes the Metropolis-Hastings ratio
}
