% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized.R
\name{penalized.logit}
\alias{penalized.logit}
\title{Penalized Logistic Regression}
\usage{
penalized.logit(X, y, beta, lambda, iter = 1000, tol = 1e-04)
}
\arguments{
\item{X}{Matrix of predictor variables.}

\item{y}{Vector of response variable.}

\item{beta}{Initial values of the coefficients.}

\item{lambda}{Vector of penalty parameters.}

\item{iter}{Maximum number of iterations (default: 1000).}

\item{tol}{Tolerance for convergence (default: 0.0001).}
}
\value{
A matrix containing the estimated coefficients for each lambda value.
}
\description{
This function performs penalized logistic regression using iterative updates
}
\details{
The function iteratively updates the coefficients using a penalized
logistic regression algorithm until convergence or maximum iterations are reached.
}
\examples{
beta_true <- c(1.54, 3, 2.3)
X <- matrix(rnorm(100), ncol = 2)
linear_predictor <- cbind(1, X) \%*\% beta_true
p <- exp(linear_predictor) / (1 + exp(linear_predictor))
y <- rbinom(50, 1, prob = p)
beta_initial <- rep(1, 3)
lambda_values <- c(0.1, 0.01, 0.001)

# Run the penalized logistic regression function
penalized.logit(X = X, y = y, beta = beta_initial, lambda = lambda_values)

}
